refactor: move timeframe related components to timeframe folder
